fix: update value to adapt for cellphones

+++ lib/shared/BackgroundObject.dart
@@ -5,7 +5,7 @@ class Bg extends SpriteComponent {
     y += speed * dt;
 
     if (y >= limit) {
-      y = posY;
+      y = posY.toDouble();
     }
   }
 }
